backend/requirements.txt
ultralytics
imutils # REDIMENSIONAR IMAGENES
paddlepaddle # FRAMEWORK DE DEEP LEARNING
paddleocr # OCR
flask
flask-cors

psycopg2-binary # Conector de PostgreSQL
asyncpg

